updated seed data
