+++ api/resolvers/index.js
@@ -1,7 +1,6 @@
 import { GraphQLDateTime } from 'graphql-iso-date';
 import { accountResolvers } from './accountResolvers';
 import { manageResolvers } from './manageResolvers';
-import { seedResolvers } from './seedResolvers';
 import { tokenResolvers } from './tokenResolvers';
 import { userResolvers } from './userResolvers';
 
@@ -14,6 +13,5 @@ export const resolvers = [
     userResolvers,
     accountResolvers,
     manageResolvers,
-    tokenResolvers,
-    seedResolvers
+    tokenResolvers
 ];
