+++ api/schema/accountSchema.js
@@ -1,4 +1,4 @@
-const { gql } = require('apollo-server');
+import { gql } from 'apollo-server';
 
 export const accountSchema = gql`
     input RegisterInput {
