+++ api/schema/index.js
@@ -1,7 +1,6 @@
 import { linkSchema } from './linkSchema';
 import { accountSchema } from './accountSchema';
 import { manageSchema } from './manageSchema';
-import { seedSchema } from './seedSchema';
 import { tokenSchema } from './tokenSchema';
 import { userSchema } from './userSchema';
 
@@ -10,6 +9,5 @@ export const typeDefs = [
     userSchema,
     accountSchema,
     manageSchema,
-    seedSchema,
     tokenSchema
 ];
