--- api/schema/seedSchema.js
@@ -1,7 +1,0 @@
-const { gql } = require('apollo-server');
-
-export const seedSchema = gql`
-    extend type Mutation {
-        seedData: MutationResponse
-    }
-`;
