+++ api/schema/tokenSchema.js
@@ -1,5 +1,4 @@
-const { gql } = require('apollo-server');
-
+import { gql } from 'apollo-server';
 export const tokenSchema = gql`
     enum GrantType {
         PASSWORD
