+++ api/schema/userSchema.js
@@ -1,4 +1,4 @@
-const { gql } = require('apollo-server');
+import { gql } from 'apollo-server';
 
 export const userSchema = gql`
     enum UserSortExpression {
